clean up macros
