+++ chain-impl-mockchain/src/ledger/tests/mod.rs
@@ -1,3 +1,5 @@
+#[macro_use]
+mod macros;
 //pub mod discrimination_tests;
 //pub mod initial_funds_tests;
 //pub mod ledger_tests;
